refactor: 컴포넌트를 expoter 파일에서 import 할 수 있도록 변경
