+++ src/assets/cartMock.json
@@ -4,8 +4,8 @@
     "quantity": 5,
     "product": {
       "id": 1,
-      "price": 10000,
-      "name": "치킨",
+      "name": "지구",
+      "price": 1000,
       "imageUrl": "https://cdn.pixabay.com/photo/2011/12/13/14/28/earth-11009__480.jpg"
     }
   },
@@ -14,8 +14,8 @@
     "quantity": 1,
     "product": {
       "id": 2,
-      "price": 20000,
-      "name": "피자",
+      "name": "화성",
+      "price": 200000,
       "imageUrl": "https://cdn.pixabay.com/photo/2011/12/13/14/30/mars-11012__480.jpg"
     }
   },
@@ -24,9 +24,9 @@
     "quantity": 1,
     "product": {
       "id": 3,
-      "price": 20000,
-      "name": "피자",
-      "imageUrl": "https://cdn.pixabay.com/photo/2011/12/13/14/30/mars-11012__480.jpg"
+      "name": "달",
+      "price": 300,
+      "imageUrl": "https://cdn.pixabay.com/photo/2016/04/02/19/40/moon-1303512__480.png"
     }
   },
   {
@@ -34,9 +34,9 @@
     "quantity": 1,
     "product": {
       "id": 4,
-      "price": 20000,
-      "name": "피자",
-      "imageUrl": "https://cdn.pixabay.com/photo/2011/12/13/14/30/mars-11012__480.jpg"
+      "name": "해왕성",
+      "price": 10000,
+      "imageUrl": "https://cdn.pixabay.com/photo/2020/09/06/22/11/neptune-5550216__480.jpg"
     }
   },
   {
@@ -44,9 +44,39 @@
     "quantity": 1,
     "product": {
       "id": 5,
-      "price": 20000,
-      "name": "피자",
-      "imageUrl": "https://cdn.pixabay.com/photo/2011/12/13/14/30/mars-11012__480.jpg"
+      "name": "태양",
+      "price": 30000000,
+      "imageUrl": "https://cdn.pixabay.com/photo/2016/07/13/20/47/sun-1515503__480.jpg"
+    }
+  },
+  {
+    "id": 6,
+    "quantity": 1,
+    "product": {
+      "id": 6,
+      "name": "달에 가려진 태양",
+      "price": 8000000,
+      "imageUrl": "https://cdn.pixabay.com/photo/2016/07/02/12/21/eclipse-1492818__480.jpg"
+    }
+  },
+  {
+    "id": 7,
+    "quantity": 1,
+    "product": {
+      "id": 7,
+      "name": "금성",
+      "price": 9990,
+      "imageUrl": "https://cdn.pixabay.com/photo/2011/12/13/14/39/venus-11022__480.jpg"
+    }
+  },
+  {
+    "id": 8,
+    "quantity": 1,
+    "product": {
+      "id": 8,
+      "name": "코스모스 별",
+      "price": 15000,
+      "imageUrl": "https://cdn.pixabay.com/photo/2018/03/08/20/01/astronomy-3209688__480.jpg"
     }
   }
 ]
