+++ src/components/Stepper.tsx
@@ -1,9 +1,10 @@
 import { ChangeEvent } from 'react';
+
 import { styled } from 'styled-components';
 
 import { useUpdateCart } from '../hooks/useUpdateCart';
 
-import { Button } from './common/Button';
+import { Button } from './exporter';
 
 import { QUANTITY } from '../constants';
